Renames store creators and adds docs
