+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "easy-peasy",
   "version": "2.4.0-alpha.4",
-  "description": "Easy peasy global state for React",
+  "description": "Easy peasy state for React",
   "license": "MIT",
   "main": "dist/easy-peasy.cjs.js",
   "module": "dist/easy-peasy.esm.js",
